mongodb_backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo[srv]==4.1.1
python-dotenv==0.19.0
gunicorn==20.1.0 